# feat: add option to disable tracing