This pull request introduces an option to significantly reduce the overhead caused by tracing instrumentation inside `Tracer` implementation in scenarios where it is not used. Metrics operations will still be performed, and their logic remains unaffected by this change. 

While otepgx already offers a method to disable tracing by providing a noop traceprovider via option, the operations that prepare trace attributes still remain. Their results will not be used, so these operations are not needed, but they still consume resources.

Here is the benchmark result demonstrating improvements:
```
goos: linux
goarch: amd64
pkg: github.com/exaring/otelpgx
cpu: Intel(R) Core(TM) i9-9900K CPU @ 3.60GHz
          │ NoopTraceProvider │      WithDisableTracing       │
          │      sec/op       │   sec/op     vs base          │
Tracer-16         740.0µ ± 2%   738.2µ ± 1%  ~ (p=0.853 n=10)

          │ NoopTraceProvider │          WithDisableTracing          │
          │       B/op        │     B/op      vs base                │
Tracer-16        3.438Ki ± 0%   1.884Ki ± 0%  -45.20% (p=0.000 n=10)

          │ NoopTraceProvider │         WithDisableTracing         │
          │     allocs/op     │ allocs/op   vs base                │
Tracer-16          65.00 ± 0%   34.00 ± 0%  -47.69% (p=0.000 n=10)
```